requirements.txt
asgiref==3.6.0
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.1.0
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.5
daphne==4.1.2
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.1
django-cors-headers==3.14.0
django-crontab==0.7.1
django-modeltranslation==0.19.3
django-oauth-toolkit==2.3.0
django-phonenumber-field==6.1.0
djangochannelsrestframework==1.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-social-oauth2==2.1.3
drf-yasg==1.21.5
ecdsa==0.18.0
geographiclib==2.0
geopy==2.3.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.5.0
MarkupSafe==2.1.2
msgpack==1.0.8
oauthlib==3.2.2
packaging==23.1
phonenumbers==8.12.45
Pillow==9.5.0
psycopg2-binary==2.9.3
pyasn1==0.5.0
pyasn1_modules==0.4.0
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==24.1.0
python-decouple==3.8
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
redis==5.0.4
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.22
ruamel.yaml.clib==0.2.7
service-identity==24.1.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
wrapt==1.15.0
zope.interface==6.3
